Fix recommended changes
